Start setting up unit tests for generators (actions)
