+++ .husky/pre-commit
@@ -4,4 +4,4 @@
 eslint .
 npx lint-staged
 npx check-outdated --ignore-packages node-plop
-# npm test
+npm test
