--- app/generators/guidance-page.js
@@ -1,55 +1,0 @@
-module.exports = function guidancePage(plop) {
-  const projectPath = process.cwd();
-
-  plop.setGenerator('guidance-page', {
-    description: 'Create text only page',
-    prompts: [{
-      type: 'input',
-      name: 'pageName',
-      message: 'Page name, include spaces we will format the name accordingly',
-    }, {
-      type: 'confirm',
-      name: 'isStartButton',
-      message: 'Do you need a start now button?',
-    }],
-
-    actions(data) {
-      const actions = [];
-
-      actions.push({
-        type: 'add',
-        path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
-        templateFile: './app/templates/common/layouts/content-and-input-page.njk',
-      });
-
-      actions.push({
-        type: 'add',
-        path: `${projectPath}/app/routes/{{kebabCase pageName}}.js`,
-        templateFile: './app/templates/guidance/guidance-controller.template.js',
-      });
-
-      actions.push({
-        type: 'modify',
-        path: `${projectPath}/app/routes.js`,
-        template: 'require(\'./routes/{{kebabCase pageName}}\')(router);\n\nmodule.exports = router',
-        pattern: /module.exports = router/gi,
-      });
-
-      actions.push({
-        type: 'modify',
-        path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
-        template: '',
-        pattern: /{# FORM #}/gi,
-      });
-
-      actions.push({
-        type: 'modify',
-        path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
-        templateFile: data.isStartButton ? './app/templates/common/components/button/start-button.njk.hbs' : './app/templates/common/components/button/continue-button.njk.hbs',
-        pattern: /{# BUTTON #}/gi,
-      });
-
-      return actions;
-    },
-  });
-};
