+++ app/generators/guidance/utils/guidance-actions.js
@@ -1,0 +1,40 @@
+module.exports = function guidanceActions(projectPath, isStartButton) {
+  const actions = [];
+
+  actions.push({
+    type: 'add',
+    path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
+    templateFile: './app/templates/common/layouts/content-and-input-page.njk',
+  });
+
+  actions.push({
+    type: 'add',
+    path: `${projectPath}/app/routes/{{kebabCase pageName}}.js`,
+    templateFile: './app/templates/guidance/guidance-controller.template.js',
+  });
+
+  actions.push({
+    type: 'modify',
+    path: `${projectPath}/app/routes.js`,
+    template: 'require(\'./routes/{{kebabCase pageName}}\')(router);\n\nmodule.exports = router',
+    pattern: /module.exports = router/gi,
+  });
+
+  actions.push({
+    type: 'modify',
+    path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
+    template: '',
+    pattern: /{# FORM #}/gi,
+  });
+
+  actions.push({
+    type: 'modify',
+    path: `${projectPath}/app/views/{{kebabCase pageName}}.njk`,
+    templateFile: isStartButton ? './app/templates/common/components/button/start-button.njk.hbs' : './app/templates/common/components/button/continue-button.njk.hbs',
+    pattern: /{# BUTTON #}/gi,
+  });
+
+  console.log(actions)
+
+  return actions;
+}
