+++ jest.config.js
@@ -1,0 +1,6 @@
+/** @type {import('jest').Config} */
+const config = {
+  modulePathIgnorePatterns: ['./govuk-prototype-kit/'],
+};
+
+module.exports = config;
